Add gitclean.sh
The script runs from the current working directory and recursively
searches for all git repos within it. For each repo, it displays the
outputs of the following commands and lets the user decide if he/she
would like to delete the repo.

  - git status
  - git branch -avv
  - git stash list

If the user chooses to delete it, the script saves a git clone command
for the deleted repo in a file called `git_deleted_repos.sh` located in
the current working directory. If the deleted repo had no remote, an
incorrect `git clone` command gets written to the file.